moving the stop words code to its own package
